[MOVE] LogoImageDeletedEvent 를 domain 으로 이동 #163
